Introduce a 'dataset' struct that represents the (image) data used during ICA
The dataset will initially be generated from the (maybe filtered) images. It represents
all the data that will be used for a singe analysis. With the introduction of the  dataset
it is possible to re-run an analysis with the exact same data.
So this can be used to look for inter analysis variation of the results.